+++ pluto-plugins/plugins/network/_base/src/main/java/com/pluto/plugins/network/internal/NetworkFragment.kt
@@ -3,4 +3,4 @@ package com.pluto.plugins.network.internal
 import androidx.fragment.app.Fragment
 import com.pluto.plugins.network.base.R
 
-class NetworkFragment : Fragment(R.layout.pluto_network___fragment_network)
+internal class NetworkFragment : Fragment(R.layout.pluto_network___fragment_network)
