+++ pluto-plugins/plugins/network/ktor-network/lib/src/main/kotlin/com/pluto/plugins/network/internal/RequestConverter.kt
@@ -1,10 +1,10 @@
-package com.pluto.plugins.network.internal.interceptor.logic.core
+package com.pluto.plugins.network.internal
 
 import com.pluto.plugins.network.internal.interceptor.logic.ApiCallData
 import com.pluto.plugins.network.internal.interceptor.logic.NetworkCallsRepo
 import com.pluto.plugins.network.internal.interceptor.logic.RequestData
 
-interface RequestConverter<T> {
+internal interface RequestConverter<T> {
     fun T.convert(): RequestData
 
     fun T.save(): ApiCallData {
