+++ pluto-plugins/plugins/network/ktor-network/lib/src/main/kotlin/com/pluto/plugins/network/internal/ResponseConverter.kt
@@ -1,10 +1,10 @@
-package com.pluto.plugins.network.internal.interceptor.logic.core
+package com.pluto.plugins.network.internal
 
 import com.pluto.plugins.network.internal.interceptor.logic.ApiCallData
 import com.pluto.plugins.network.internal.interceptor.logic.NetworkCallsRepo
 import com.pluto.plugins.network.internal.interceptor.logic.ResponseData
 
-interface ResponseConverter<T> {
+internal interface ResponseConverter<T> {
     suspend fun T.convert(): ResponseData
 
     suspend fun T.save(apiCallData: ApiCallData) {
