+++ sample/src/main/java/com/sampleapp/functions/network/internal/core/KtorNetwork.kt
@@ -1,6 +1,6 @@
 package com.sampleapp.functions.network.internal.core
 
-import com.pluto.plugins.network.ktor.addPlutoKtorPlugin
+import com.pluto.plugins.network.addPlutoKtorPlugin
 import io.ktor.client.HttpClient
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.client.plugins.defaultRequest
